lib/security/pyproject.toml
[project]
name = "security"
version = "0.1.0"
description = "Security scanning for Codex projects"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "cooneycw"}
]
dependencies = []

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
all = ["pyyaml>=6.0"]

[project.scripts]
security-scan = "security.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
